create new nfc session when old one is invalidated.
According to Apple's documentation, when a reader session is invalidated,
it is closed and subsequently no longer reusable:

https://developer.apple.com/documentation/corenfc/nfcreadersessionprotocol/2874107-invalidate

This is the cause of #1 in the `User Experiences` remarks.

```
Scanning an NDEF-tag usually works once directly after rebooting the iPhone. From then on, it may or may not work, usually it doesn't work and another reboot is required. This was seen with Beta 1-4 of iOS 11.
```

My workaround (which I think is the expected use) is to initialize a new
session when the other one is invalidated.

Now you can scan as many times as you'd like without having to quit the app, yay!